# src/main/resources/geojson-service-api-docs.json
{"openapi":"3.0.1","info":{"title":"Pixup GeoJSON Microservice Endpoints","description":"Definición de los Endpoints de GeoJSON Service para el sistema Pixup","contact":{"name":"GIS","url":"https://www.unam.mx/","email":"jlvillarreal.benitez@gmail.com.mx"},"version":"v1"},"servers":[{"url":"http://localhost:8082","description":"Generated server url"}],"tags":[{"name":"localidad","description":"API del Recurso Localidad"}],"paths":{"/api/reservas/{id}":{"get":{"tags":["reserva-controller"],"operationId":"reservaPorId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Reserva"}}}}}},"put":{"tags":["reserva-controller"],"operationId":"actualizarReserva","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Reserva"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Reserva"}}}}}},"delete":{"tags":["reserva-controller"],"operationId":"eliminarReserva","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK"}}}},"/api/localidades/registro":{"post":{"tags":["localidad"],"summary":"Registrar Localidad","operationId":"registrarLocalidad","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistroLocalidadRequest"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"409":{"description":"Ya existe una localidad con ese número de reporte","content":{}},"201":{"description":"Localidad Creada Exitosamente","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Localidad"}}}}}}},"/api/reservas":{"get":{"tags":["reserva-controller"],"operationId":"obtenerReservas","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}}},"post":{"tags":["reserva-controller"],"operationId":"crearReserva","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Reserva"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Reserva"}}}}}}}},"components":{"schemas":{"Reserva":{"type":"object","properties":{"id":{"type":"string"},"nombre":{"type":"string"}}},"Zona":{"required":["tipo","cantidad"],"type":"object","properties":{"tipo":{"maxLength":60,"minLength":3,"type":"string"},"cantidad":{"maxLength":30,"minLength":1,"type":"string"},"numero":{"type":"string"}}},"RegistroLocalidadRequest":{"required":["zona","localidad"],"type":"object","properties":{"localidad":{"$ref":"#/components/schema/Localidad"},"zona":{"$ref":"#/components/schemas/Zona"}}},"Localidad":{"required":["email","nombre"],"type":"object","properties":{"id":{"type":"string"},"nombre":{"type":"string"},"coorX":{"type":"string"},"coorY":{"type":"string"},"altitud":{"type":"string"},"numReporte":{"type":"string"},"tipoReporte":{"type":"string"},"zonas":{"type":"array","items":{"$ref":"#/components/schemas/Zona"}}}}}}}